feat: support providing LDES_VERSION_OF_PATH and TIMESTAMP_PATH
If the consuming feed doesn't provide the metadata users of the service
can still define the expected paths and this service will use them.

Note that `REPLACE_VERSIONS` will not work when providing these
because the ldes-client library needs the feed to provide the paths itself,
we can't tell it what paths to look for. An alternative would be to
re-implement versioning in this service and fallback to it if the feed
doesn't provide the paths.